Force to __init__
