+++ homeassistant/components/accuweather/__init__.py
@@ -2,11 +2,13 @@
 
 from __future__ import annotations
 
+from dataclasses import dataclass
 import logging
 
 from accuweather import AccuWeather
 
 from homeassistant.components.sensor import DOMAIN as SENSOR_PLATFORM
+from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_API_KEY, CONF_NAME, Platform
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import entity_registry as er
@@ -14,15 +16,24 @@ from homeassistant.helpers.aiohttp_client import async_get_clientsession
 
 from .const import DOMAIN, UPDATE_INTERVAL_DAILY_FORECAST, UPDATE_INTERVAL_OBSERVATION
 from .coordinator import (
-    AccuWeatherConfigEntry,
     AccuWeatherDailyForecastDataUpdateCoordinator,
-    AccuWeatherData,
     AccuWeatherObservationDataUpdateCoordinator,
 )
 
 _LOGGER = logging.getLogger(__name__)
 
 PLATFORMS = [Platform.SENSOR, Platform.WEATHER]
+
+
+@dataclass
+class AccuWeatherData:
+    """Data for AccuWeather integration."""
+
+    coordinator_observation: AccuWeatherObservationDataUpdateCoordinator
+    coordinator_daily_forecast: AccuWeatherDailyForecastDataUpdateCoordinator
+
+
+type AccuWeatherConfigEntry = ConfigEntry[AccuWeatherData]
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: AccuWeatherConfigEntry) -> bool:
