+++ homeassistant/components/airnow/__init__.py
@@ -16,10 +16,11 @@ from homeassistant.helpers import device_registry as dr, entity_registry as er
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 
 from .coordinator import AirNowDataUpdateCoordinator
-from .types import AirNowConfigEntry
 
 _LOGGER = logging.getLogger(__name__)
 PLATFORMS = [Platform.SENSOR]
+
+type AirNowConfigEntry = ConfigEntry[AirNowDataUpdateCoordinator]
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: AirNowConfigEntry) -> bool:
